Refactor PatternVerifier
A separate module is split off for the PatternVerifier definition, to avoid
module cycles in the future. The builtin pattern verifiers are also separated
from the builtin function evaluators.
