+++ kore/src/Kore/Builtin/Endianness.hs
@@ -51,8 +51,8 @@ import qualified Kore.Verified as Verified
 verifiers :: Verifiers
 verifiers =
     mempty
-        { patternVerifier =
-            (applicationPatternVerifiers . HashMap.fromList)
+        { patternVerifierHook =
+            (applicationPatternVerifierHooks . HashMap.fromList)
                 [ (KlabelSymbolKey littleEndianKey, littleEndianVerifier)
                 , (KlabelSymbolKey bigEndianKey   , bigEndianVerifier   )
                 ]
