+++ kore/src/Kore/Builtin/Int.hs
@@ -113,7 +113,7 @@ verifiers =
     Builtin.Verifiers
         { sortDeclVerifiers
         , symbolVerifiers
-        , patternVerifier
+        , patternVerifierHook
         }
 
 {- | Verify that hooked sort declarations are well-formed.
@@ -175,9 +175,9 @@ symbolVerifiers =
 
 {- | Verify that domain value patterns are well-formed.
  -}
-patternVerifier :: Builtin.PatternVerifier (TermLike variable)
-patternVerifier =
-    Builtin.domainValuePatternVerifier sort
+patternVerifierHook :: Builtin.PatternVerifierHook
+patternVerifierHook =
+    Builtin.domainValuePatternVerifierHook sort
     $ Builtin.makeEncodedDomainValueVerifier sort patternVerifierWorker
   where
     patternVerifierWorker external =
