+++ kore/src/Kore/Builtin/InternalBytes.hs
@@ -72,7 +72,7 @@ verifiers =
     Builtin.Verifiers
         { sortDeclVerifiers
         , symbolVerifiers
-        , patternVerifier
+        , patternVerifierHook
         }
 
 -- | Verify that hooked sort declarations are well-formed.
@@ -133,9 +133,9 @@ symbolVerifiers =
 
 {- | Verify that domain value patterns are well-formed.
  -}
-patternVerifier :: Builtin.PatternVerifier (TermLike variable)
-patternVerifier =
-    Builtin.domainValuePatternVerifier sort
+patternVerifierHook :: Builtin.PatternVerifierHook
+patternVerifierHook =
+    Builtin.domainValuePatternVerifierHook sort
     $ Builtin.makeEncodedDomainValueVerifier sort patternVerifierWorker
   where
     patternVerifierWorker external =
