+++ kore/src/Kore/Builtin/KEqual.hs
@@ -62,7 +62,7 @@ verifiers =
     Builtin.Verifiers
         { sortDeclVerifiers = mempty
         , symbolVerifiers
-        , patternVerifier = mempty
+        , patternVerifierHook = mempty
         }
 
 {- | Verify that hooked symbol declarations are well-formed.
