+++ kore/src/Kore/Builtin/List.hs
@@ -126,7 +126,7 @@ verifiers =
     Builtin.Verifiers
         { sortDeclVerifiers
         , symbolVerifiers
-        , patternVerifier = mempty
+        , patternVerifierHook = mempty
         }
 
 {- | Verify that hooked sort declarations are well-formed.
