+++ kore/src/Kore/Builtin/Map.hs
@@ -117,7 +117,7 @@ verifiers =
     Builtin.Verifiers
         { sortDeclVerifiers
         , symbolVerifiers
-        , patternVerifier = mempty
+        , patternVerifierHook = mempty
         }
 
 {- | Verify that hooked sort declarations are well-formed.
